Finished testing first transition algorithm. Building the AST list and tables for labels is also tested and done. Everything passed valgrind tests.
Made many changes to various files on the project, including:
everything related to the first transition algorithm;
FileHandling, NameTable, and AST interfaces to pass valgrind checks and be relevant more to the project.

Decided to finish with the second transition algorithm before implementing the error checks and handling.

Also, need to add documentation to first transition functions, and to split the AST interface to multiple files (not mandatory).

(Monday, 07/08/23)